+++ src/UndefTemplate.ts
@@ -1,7 +1,0 @@
-/**
- * Utility type to include the possibility of an object to be undefined.
- *
- * @typedef {UndefTemplate<T>}
- * @template T
- */
-export type UndefTemplate<T> = T | undefined;