+++ src/test/suite/index.ts
@@ -21,13 +21,13 @@ export function run(): Promise<void> {
       }
 
       // Add files to the test suite
-      files.forEach((f) => mocha.addFile(path.resolve(testsRoot, f)));
+      files.forEach(file => mocha.addFile(path.resolve(testsRoot, file)));
 
       try {
         // Run the mocha test
-        mocha.run((failures) => {
+        mocha.run(failures => {
           if(failures > 0) {
-            e(new Error(failures + 'tests failed.'));
+            e(new Error(`${failures}  tests failed.`));
           } else {
             c();
           }