+++ CHANGELOG.md
@@ -6,22 +6,37 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 Nothing new
 
-## [1.1.1] - 2022/04/09
+## [1.2.0] - 2023/03/dd
+
 ### Added
+
+- Setting to toggle whether to include types.
+- Support for JavaScript and JavaScript React.
+
+## [1.1.1] - 2022/04/09
+
+### Added
+
 - Setting to toggle whether to include export tag
 - Setting to toggle whether to include async tag
 
 ## [1.1.0] - 2021/12/03
+
 ### Added
+
 - Support for TypeScript React.
+
 ### Changed
+
 - A few minor project configuration settings with consequent fixes.
 
-
 ## [1.0.0] - 2021/07/20
+
 ### Added
+
 - 3 commands declarations.
 - Extension configurations.
 - Bundling.
@@ -33,9 +48,10 @@ Nothing new
 - Few basic unit tests.
 - This [CHANGELOG] file.
 
-[Unreleased]: https://github.com/Nyphet/jsdoc-generator
+[unreleased]: https://github.com/Nyphet/jsdoc-generator
+[1.2.0]: https://github.com/Nyphet/jsdoc-generator/releases/tag/v1.2.0
 [1.1.1]: https://github.com/Nyphet/jsdoc-generator/releases/tag/v1.1.1
 [1.1.0]: https://github.com/Nyphet/jsdoc-generator/releases/tag/v1.1.0
 [1.0.0]: https://github.com/Nyphet/jsdoc-generator/releases/tag/v1.0.0
-[README]: https://github.com/Nyphet/jsdoc-generator
-[CHANGELOG]: https://github.com/Nyphet/jsdoc-generator/blob/main/CHANGELOG.md
+[readme]: https://github.com/Nyphet/jsdoc-generator
+[changelog]: https://github.com/Nyphet/jsdoc-generator/blob/main/CHANGELOG.md